Add responsiveness to buttons, styling
Kinda forgot to commit for a while so a bunch of code just appeared out
of nowhere, oops. Made the function that maps out where exactly the user
is clicking on the screen, and returns its coordinates, which should be
very useful in modifying the array once we get to that point. Gotta flesh
out the gamechange function to include the player factory objects as a
reference rather than just a string.