Fix game piece assignment logic
I was struggling with how to select the game pieces, and give them
a unique identifier. I hacked out a coordinate system using a variable
that looped in the first loop of row creation, but I could have it loop
inside of the individual cell creation, and actually give them all
unique ids which we can use to select them now. Let's go.