Add wincon logic, cleaned code
Uh so in this iteration, we figured out the player factory.
and now you can have your very own unique marker and player name,
it'll reset the board for you once you win, and then uh it'll give the
loser the starting position (as a result of me not changing the game
state on reset(could be a secret feature haha).) should be simple enough
in the future to reset fully and add new players, but for now I'm happy
with this.